Fixed Broken Link
Note: the link seems to redirect to https://www.richdevelops.dev/blog/building-apps-with-sam-typescript-and-vs-code-debugging. In the spirit of minimal change though, all that seemed to be required was adding the dash between "vs" and "code". I also updated the protocol to https for good measure.

Perhaps though the link should just go directly to https://www.richdevelops.dev/blog/building-apps-with-sam-typescript-and-vs-code-debugging?