Example of Passing Options
It wasn't clear to me immediately how the options should be passed. I initially thought they would maybe go in as command line options. But then I discovered the constructor parameter. I think it's good to have an example so people can see how the options are used. If this is not the only way to pass options, then it might be good to have an example of other ways to pass them too.